CardScore scores left and right bowers
I'm breaking the ping pong rules because I made some decisions I'm not sure about and I think there is an opportunity to refactor. I think that should be done as a pair, so pushing this to get some feedback.

I made Suit a real thing, but it feels a little early. I'm wondering if it should go back to just a symbol, and then the concept of "same color suits" could be a private method on CardScore. Alternatively, the check in left_bower? where it dives into Suit to check its color could move to Suit. I'm just not sure which direction to go in yet. Thoughts?